rest endpoint consistency
