+++ crates/evops-rest/src/routes/v1/events.rs
@@ -7,20 +7,22 @@ use axum::extract::State;
 use crate::error::AddResponse as _;
 
 mod _id;
-mod create;
 
 fn route_docs(r: TransformPathItem) -> TransformPathItem {
     r.tag(crate::docs::Tag::EventService.into())
 }
 pub fn router() -> ApiRouter<crate::AppState> {
     ApiRouter::new()
-        .api_route_with("/", get_with(self::get, self::get_docs), self::route_docs)
+        .api_route_with(
+            "/",
+            get_with(self::get, self::get_docs).post_with(self::post, self::post_docs),
+            self::route_docs,
+        )
         .nest("/{id}", self::_id::router())
-        .nest("/create", self::create::router())
 }
 
 fn get_docs(o: TransformOperation) -> TransformOperation {
-    o.summary("evops.api.v1.EventService/List")
+    o.summary("evops.api.v1.EventService.List")
         .response_bad_request() // TODO:
         .response_unprocessable_entity()
         .response_internal_server_error()
@@ -36,4 +38,17 @@ async fn get(
             .await?
             .into()
     }))
+}
+
+fn post_docs(o: TransformOperation) -> TransformOperation {
+    o.summary("evops.api.v1.EventService.Create")
+        .response_bad_request()
+        .response_unprocessable_entity()
+        .response_internal_server_error()
+}
+async fn post(
+    State(state): State<crate::AppState>,
+    Json(request): Json<crate::types::EventServiceCreateRequest>,
+) -> crate::Result<Json<crate::types::EventServiceCreateResponse>> {
+    Ok(Json(state.create_event(request.try_into()?).await?.into()))
 }
