--- crates/evops-rest/src/routes/v1/tags/create.rs
@@ -1,32 +1,0 @@
-use aide::axum::ApiRouter;
-use aide::axum::routing::post_with;
-use aide::transform::{TransformOperation, TransformPathItem};
-use axum::Json;
-use axum::extract::State;
-
-use crate::error::AddResponse as _;
-
-fn route_docs(r: TransformPathItem) -> TransformPathItem {
-    r.tag(crate::docs::Tag::TagService.into())
-}
-pub fn router() -> ApiRouter<crate::AppState> {
-    ApiRouter::new().api_route_with(
-        "/",
-        post_with(self::post, self::post_docs),
-        self::route_docs,
-    )
-}
-
-fn post_docs(o: TransformOperation) -> TransformOperation {
-    o.summary("evops.api.v1.TagService/Create")
-        .response_bad_request()
-        .response_conflict()
-        .response_unprocessable_entity()
-        .response_internal_server_error()
-}
-async fn post(
-    State(state): State<crate::AppState>,
-    Json(request): Json<crate::types::TagServiceCreateRequest>,
-) -> crate::Result<Json<crate::types::TagServiceCreateResponse>> {
-    Ok(Json(state.create_tag(request.try_into()?).await?.into()))
-}
